+++ api/.snapshots/TestUserAPI-Queries-Get_user_list_using_ACTIVE_DESC
@@ -19,6 +19,14 @@
                 ]
             },
             {
+                "id": "dXNlcjoz",
+                "email": "user3@example.com",
+                "createdAt": "2007-07-07T12:00:00Z",
+                "updatedAt": "2007-07-07T12:00:00Z",
+                "isActive": true,
+                "pages": []
+            },
+            {
                 "id": "dXNlcjoy",
                 "email": "user2@example.com",
                 "createdAt": "2007-07-07T11:00:00Z",
@@ -30,14 +38,6 @@
                         "name": "Page 3"
                     }
                 ]
-            },
-            {
-                "id": "dXNlcjoz",
-                "email": "user3@example.com",
-                "createdAt": "2007-07-07T12:00:00Z",
-                "updatedAt": "2007-07-07T12:00:00Z",
-                "isActive": true,
-                "pages": []
             },
             {
                 "id": "dXNlcjo0",
