+++ api/.snapshots/TestUserAPI-Queries-Get_user_list_using_EMAIL_DESC
@@ -1,0 +1,52 @@
+{
+    "data": {
+        "users": [
+            {
+                "id": "dXNlcjo0",
+                "email": "user4@example.com",
+                "createdAt": "2007-07-07T13:00:00Z",
+                "updatedAt": "2007-07-07T13:00:00Z",
+                "isActive": false,
+                "pages": []
+            },
+            {
+                "id": "dXNlcjoz",
+                "email": "user3@example.com",
+                "createdAt": "2007-07-07T12:00:00Z",
+                "updatedAt": "2007-07-07T12:00:00Z",
+                "isActive": true,
+                "pages": []
+            },
+            {
+                "id": "dXNlcjoy",
+                "email": "user2@example.com",
+                "createdAt": "2007-07-07T11:00:00Z",
+                "updatedAt": "2007-07-07T11:00:00Z",
+                "isActive": false,
+                "pages": [
+                    {
+                        "id": "cGFnZToz",
+                        "name": "Page 3"
+                    }
+                ]
+            },
+            {
+                "id": "dXNlcjox",
+                "email": "user1@example.com",
+                "createdAt": "2007-07-07T10:00:00Z",
+                "updatedAt": "2007-07-07T10:00:00Z",
+                "isActive": true,
+                "pages": [
+                    {
+                        "id": "cGFnZTox",
+                        "name": "Page 1"
+                    },
+                    {
+                        "id": "cGFnZToy",
+                        "name": "Page 2"
+                    }
+                ]
+            }
+        ]
+    }
+}
