+++ api/schema/input/UserSort.graphql
@@ -1,0 +1,4 @@
+input UserSort {
+    field: UserSortableField!
+    order: SortOrder!
+}
